refactor: remove old chat panel provider
