+++ vscode/src/chat/chat-view/ChatManager.ts
@@ -16,24 +16,24 @@ import { CachedRemoteEmbeddingsClient } from '../CachedRemoteEmbeddingsClient'
 import { AuthStatus } from '../protocol'
 
 import { ChatPanelsManager, IChatPanelProvider } from './ChatPanelsManager'
-import { SidebarChatOptions, SidebarChatProvider } from './SidebarChatProvider'
+import { SidebarViewOptions, SidebarViewProvider } from './SidebarViewProvider'
 
 export const CodyChatPanelViewType = 'cody.chatPanel'
 /**
  * Manages chat view providers and panels.
  */
 export class ChatManager implements vscode.Disposable {
-    // View in sidebar for auth flow and old chat sidebar view
+    // SidebarViwq is used for auth view and running tasks that do not require a chat view
     // We will always keep an instance of this around (even when not visible) to handle states when no panels are open
-    public sidebarChat: SidebarChatProvider
+    public sidebarView: SidebarViewProvider
     private chatPanelsManager: ChatPanelsManager
 
-    private options: SidebarChatOptions
+    private options: SidebarViewOptions
 
     protected disposables: vscode.Disposable[] = []
 
     constructor(
-        { extensionUri, ...options }: SidebarChatOptions,
+        { extensionUri, ...options }: SidebarViewOptions,
         private chatClient: ChatClient,
         private embeddingsClient: CachedRemoteEmbeddingsClient,
         private localEmbeddings: LocalEmbeddingsController | null,
@@ -46,7 +46,7 @@ export class ChatManager implements vscode.Disposable {
         )
         this.options = { extensionUri, ...options }
 
-        this.sidebarChat = new SidebarChatProvider(this.options)
+        this.sidebarView = new SidebarViewProvider(this.options)
 
         this.chatPanelsManager = new ChatPanelsManager(
             this.options,
@@ -86,7 +86,7 @@ export class ChatManager implements vscode.Disposable {
     }
 
     /**
-     * Executes a recipe in the chat view.
+     * Executes a recipe in a view provider.
      */
     public async executeRecipe(
         recipeId: RecipeID,
@@ -100,20 +100,15 @@ export class ChatManager implements vscode.Disposable {
             return
         }
 
-        // If chat view is not needed, run the recipe via sidebar chat without creating a new panel
+        // If chat view is not needed, run the recipe via sidebar view without creating a new panel
         const isDefaultEditCommands = ['/doc', '/edit'].includes(humanChatInput)
         if (!openChatView || isDefaultEditCommands) {
-            await this.sidebarChat.executeRecipe(recipeId, humanChatInput, source)
+            await this.sidebarView.executeRecipe(recipeId, humanChatInput, source)
             return
         }
 
         // Else, open a new chanel panel and run the command in the new panel
         const chatProvider = await this.getChatProvider()
-        if (!openChatView || !this.chatPanelsManager) {
-            await this.sidebarChat.executeRecipe(recipeId, humanChatInput, source)
-            return
-        }
-
         await chatProvider.executeRecipe(recipeId, humanChatInput, source)
     }
 
@@ -133,15 +128,14 @@ export class ChatManager implements vscode.Disposable {
         const chatID = treeItem?.id
         const chatLabel = treeItem?.label as vscode.TreeItemLabel
         if (chatID) {
-            await this.chatPanelsManager?.editChatHistory(chatID, chatLabel.label)
+            await this.chatPanelsManager.editChatHistory(chatID, chatLabel.label)
         }
     }
 
     public async clearHistory(treeItem?: vscode.TreeItem): Promise<void> {
         const chatID = treeItem?.id
         if (chatID) {
-            await this.sidebarChat.clearChatHistory(chatID)
-            await this.chatPanelsManager?.clearHistory(chatID)
+            await this.chatPanelsManager.clearHistory(chatID)
             return
         }
 
@@ -158,8 +152,7 @@ export class ChatManager implements vscode.Disposable {
                 return
             }
 
-            await this.sidebarChat.clearHistory()
-            await this.chatPanelsManager?.clearHistory()
+            await this.chatPanelsManager.clearHistory()
         }
     }
 
@@ -180,11 +173,11 @@ export class ChatManager implements vscode.Disposable {
      */
     public async exportHistory(): Promise<void> {
         // Use sidebar chat view for non-chat-session specfic actions
-        await this.sidebarChat.exportHistory()
+        await this.sidebarView.exportHistory()
     }
 
     public async simplifiedOnboardingReloadEmbeddingsState(): Promise<void> {
-        await this.sidebarChat.simplifiedOnboardingReloadEmbeddingsState()
+        await this.sidebarView.simplifiedOnboardingReloadEmbeddingsState()
     }
 
     /**
@@ -235,15 +228,15 @@ export class ChatManager implements vscode.Disposable {
     }
 
     private disposeChatPanelsManager(): void {
-        this.options.contextProvider.webview = this.sidebarChat.webview
-        this.chatPanelsManager?.dispose()
+        this.options.contextProvider.webview = this.sidebarView.webview
+        this.chatPanelsManager.dispose()
     }
 
     // For registering the commands for chat panels in advance
     private async createNewWebviewPanel(): Promise<void> {
         const debounceCreatePanel = debounce(
             async () => {
-                await this.chatPanelsManager?.createWebviewPanel()
+                await this.chatPanelsManager.createWebviewPanel()
             },
             250,
             { leading: true, trailing: true }
@@ -257,7 +250,7 @@ export class ChatManager implements vscode.Disposable {
     private async restorePanel(chatID: string, chatQuestion?: string): Promise<void> {
         const debounceRestore = debounce(
             async (chatID: string, chatQuestion?: string) => {
-                await this.chatPanelsManager?.restorePanel(chatID, chatQuestion)
+                await this.chatPanelsManager.restorePanel(chatID, chatQuestion)
             },
             250,
             { leading: true, trailing: true }
